requirements.txt
asgiref==3.2.8
astroid==2.4.2
Django==3.0.7
django-uuslug==1.2.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
Pillow==7.1.2
pylint==2.5.3
python-slugify==4.0.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
wrapt==1.12.1
